+++ src/test/java/com/evanzeimet/wirecucumber/TestWireCucumber.java
@@ -5,6 +5,14 @@ import com.evanzeimet.wirecucumber.scenario.TestSteps;
 public class TestWireCucumber
 		extends WireCucumber {
 
+	public TestWireCucumber() {
+		super();
+	}
+
+	public TestWireCucumber(WireCucumberOptions options) {
+		super(options);
+	}
+
 	protected TestSteps getTestSteps() {
 		return new TestSteps(steps);
 	}